requirements.txt
Flask==1.1.2
Flask-Bootstrap==3.3.7.1
Flask-Datepicker==0.14
Flask-Environments==0.1
Flask-Session==0.3.2
flask-redis==0.4.0
Flask-Login==0.5.0
Flask-WTF==0.14.3
geopy==2.0.0
idna==2.10
iniconfig==1.1.1
itsdangerous==1.1.0
Jinja2==2.11.2
kombu==5.0.2
Mako==1.1.3
MarkupSafe==1.1.1
packaging==20.4
pip-tools==5.3.1
pluggy==0.13.1
prompt-toolkit==3.0.8
py==1.9.0
Pygments==2.7.2
pyparsing==2.4.7
python-dateutil==2.8.1
python-editor==1.0.4
pytz==2020.1
PyYAML==5.3.1
redis==3.5.3
requests==2.24.0
six==1.15.0
text-unidecode==1.3
toml==0.10.1
ujson==1.35
urllib3==1.25.11
vine==5.0.0
visitor==0.1.3
wcwidth==0.2.5
Werkzeug==1.0.1
WTForms==2.3.3
timeago==1.0.14
dnspython==2.0.0
email-validator==1.1.1
Pillow==8.4.0